bridge: fix edge case of invisible wall on bridge.
Also convert all height code to use int32_t like Tomb5.
Also fixes the need for back and forth changes between int16_t and int32_t.

+++ src/game/control_util.c
@@ -165,7 +165,7 @@ int16_t GetWaterHeight(int32_t x, int32_t y, int32_t z, int16_t room_num)
     }
 }
 
-int16_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
+int32_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
 {
     g_HeightType = HT_WALL;
     while (floor->pit_room != NO_ROOM) {
@@ -175,7 +175,7 @@ int16_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
         floor = &r->floor[x_floor + y_floor * r->x_size];
     }
 
-    int16_t height = floor->floor << 8;
+    int32_t height = floor->floor << 8;
 
     g_TriggerIndex = NULL;
 
@@ -191,8 +191,8 @@ int16_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
 
         switch (type & DATA_TYPE) {
         case FT_TILT: {
-            int16_t xoff = data[0] >> 8;
-            int16_t yoff = (int8_t)data[0];
+            int32_t xoff = data[0] >> 8;
+            int32_t yoff = (int8_t)data[0];
 
             if (!g_ChunkyFlag || (ABS(xoff) <= 2 && ABS(yoff) <= 2)) {
                 if (ABS(xoff) > 2 || ABS(yoff) > 2) {
@@ -202,17 +202,15 @@ int16_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
                 }
 
                 if (xoff < 0) {
-                    height -= (int16_t)((xoff * (z & (WALL_L - 1))) >> 2);
+                    height -= ((xoff * (z & (WALL_L - 1))) >> 2);
                 } else {
-                    height +=
-                        (int16_t)((xoff * ((WALL_L - 1 - z) & (WALL_L - 1))) >> 2);
+                    height += ((xoff * ((WALL_L - 1 - z) & (WALL_L - 1))) >> 2);
                 }
 
                 if (yoff < 0) {
-                    height -= (int16_t)((yoff * (x & (WALL_L - 1))) >> 2);
+                    height -= ((yoff * (x & (WALL_L - 1))) >> 2);
                 } else {
-                    height +=
-                        (int16_t)((yoff * ((WALL_L - 1 - x) & (WALL_L - 1))) >> 2);
+                    height += ((yoff * ((WALL_L - 1 - x) & (WALL_L - 1))) >> 2);
                 }
             }
 
@@ -260,7 +258,7 @@ int16_t GetHeight(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
     return height;
 }
 
-int16_t GetCeiling(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
+int32_t GetCeiling(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
 {
     int16_t *data;
     int16_t type;
@@ -274,7 +272,7 @@ int16_t GetCeiling(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
         f = &r->floor[x_floor + y_floor * r->x_size];
     }
 
-    int16_t height = f->ceiling << 8;
+    int32_t height = f->ceiling << 8;
 
     if (f->index) {
         data = &g_FloorData[f->index];
@@ -292,17 +290,15 @@ int16_t GetCeiling(FLOOR_INFO *floor, int32_t x, int32_t y, int32_t z)
             if (!g_ChunkyFlag
                 || (xoff >= -2 && xoff <= 2 && yoff >= -2 && yoff <= 2)) {
                 if (xoff < 0) {
-                    height += (int16_t)((xoff * (z & (WALL_L - 1))) >> 2);
+                    height += ((xoff * (z & (WALL_L - 1))) >> 2);
                 } else {
-                    height -=
-                        (int16_t)((xoff * ((WALL_L - 1 - z) & (WALL_L - 1))) >> 2);
+                    height -= ((xoff * ((WALL_L - 1 - z) & (WALL_L - 1))) >> 2);
                 }
 
                 if (yoff < 0) {
-                    height +=
-                        (int16_t)((yoff * ((WALL_L - 1 - x) & (WALL_L - 1))) >> 2);
+                    height += ((yoff * ((WALL_L - 1 - x) & (WALL_L - 1))) >> 2);
                 } else {
-                    height -= (int16_t)((yoff * (x & (WALL_L - 1))) >> 2);
+                    height -= ((yoff * (x & (WALL_L - 1))) >> 2);
                 }
             }
         }
