+++ src/game/objects/trapdoor.c
@@ -26,7 +26,7 @@ void TrapDoorControl(int16_t item_num)
 }
 
 void TrapDoorFloor(
-    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int16_t *height)
+    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int32_t *height)
 {
     if (!OnTrapDoor(item, x, z)) {
         return;
@@ -38,7 +38,7 @@ void TrapDoorFloor(
 }
 
 void TrapDoorCeiling(
-    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int16_t *height)
+    ITEM_INFO *item, int32_t x, int32_t y, int32_t z, int32_t *height)
 {
     if (!OnTrapDoor(item, x, z)) {
         return;
